fixed #2269;
