+++ .gitignore
@@ -42,5 +42,5 @@ tmp
 package-lock.json
 /packages/*/coverage/**/*
 
-testCafe/screenshots
-testCafe/artifacts
+visualRegressionTests/screenshots
+visualRegressionTests/artifacts
