+++ packages/survey-creator-react/package.json
@@ -19,10 +19,12 @@
     "test_debug": "node --inspect-brk ./node_modules/jest/bin/jest.js --runInBand",
     "doc:gen": "node doc_generator/editor_docgenerator.js src/entries/index.ts",
     "doc:update": "chmod +x docupdate_npm.sh && ./docupdate_npm.sh",
-    "testcafe": "concurrently \"http-server --silent\" \"testcafe -c 4 chrome ../../testCafe/ -s ../../ --selector-timeout 1500\"",
-    "testcafe:ci": "http-server --silent & testcafe chrome:headless ../../testCafe/ -s ../../ --selector-timeout 1500 --reporter minimal,dashboard",
-    "testcafe:file": "concurrently \"http-server --silent\" \"testcafe chrome ../../testCafe/designer/toolbox.ts -s ../../ --reporter minimal --debug-mode\"",
-    "testcafe:a11y": "concurrently \"http-server --silent\" \"testcafe chrome ./testCafe/accessibility.js -s ../../ --selector-timeout 1500\"",
+    "testcafe": "concurrently \"http-server --silent\" \"testcafe -c 4 chrome ../../testCafe/ --selector-timeout 1500\"",
+    "testcafe:ci": "http-server --silent & testcafe chrome:headless ../../testCafe/ --selector-timeout 1500 --reporter minimal,dashboard",
+    "testcafe:file": "concurrently \"http-server --silent\" \"testcafe chrome ../../testCafe/designer/toolbox.ts --reporter minimal --debug-mode\"",
+    "testcafe:a11y": "concurrently \"http-server --silent\" \"testcafe chrome ./testCafe/accessibility.js --selector-timeout 1500\"",
+    "visual-regression-tests": "http-server --silent & testcafe chrome:headless ../../visualRegressionTests/ --screenshots ../../ --selector-timeout 1500 --reporter minimal,dashboard",
+    "visual-regression-tests:file": "concurrently \"http-server --silent\" \"testcafe chrome ../../visualRegressionTests/tests/designer/toolbox.ts --screenshots ../../ --reporter minimal --selector-timeout 1500\"",
     "release": "standard-version --message \"Release: %s [skip ci]\" --prerelease beta",
     "remove-package-lock": "rimraf package-lock.json",
     "build": "webpack --env.buildType dev && webpack --env.buildType prod",
