+++ visualRegressionTests/tests/designer/etalons/toolbox.png
